update file namings
